resume-extractor/requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
python-docx==1.1.0
PyPDF2==3.0.1
pandas==2.2.1
python-dotenv==1.0.1 